Move detection of anonymous struct or union into parseVarSeq().
The detection is in parseVar(). This function is called to parse the
top-level of the local variable dump, the printed watched expressions, and
the contents of structures (recursively). Only the latter one can contain
detectable anonymous struct or union members. The former expand the members
as top-level variables. Move the detection to the only place where it is
needed.

This prepares for a later change where we want to detect anonymous struct
or union members that occur as the first member of a structure.