Convert ExprValue's child nodes to a std::list<>.
ExprValue represents the value of a variable including compound values. For
these, the member values are held in a singly linked list anchored at
the m_child member.

We will want to iterate over the child nodes with std algorithms, but this
simple data structure does not allow to do that. Convert the child node
container to a std::list<ExprValue*>.

Note that with this change the recursion depth in the ExprValue destructor
no longer depends on the number of children, but only on the depth of the
value tree.