Recognize anonymous members even if they are the first member.
Our parser can recognize anonymous unions only when they are not the first
member of the enclosing structure, because to parse nested values, it
determines in advance whether it is dealing with a structure with named
members or an array, which is just a list of values.

The cue that there is not an array is the presence of a name followed by
an equal sign at the very beginning of the nested structure. However, if
the first member of the structure is anonymous, then the cue is absent and
the text looks like an array. The parser proceeds under this assumption,
but later stumbles over the "NAME =" of the next regular member.

Rewrite the parser in the following way:

- Do not decide early which kind of nested structure to parse.

- Merge the bodies of the two algorithms such that the presence of a name
  is determined for each member separately.

- Distinguish between array and structure after having parsed all nested
  values. If any of the members has a name, the container is a structure,
  otherwise, it is an array.

- Format array indices only after it is known that it is an array and mark
  unnamed members of a structure as anonymous.

With this strategy, anonymous structs or unions are detected even if they
occur as the first member of the containing struct.