Distinguish structs that contain only anonymous structs from arrays.
GDB prints a struct that has only anonymous members in a form that looks
like an array, because no immediate member value has a name. If any two of
the members have members with different names (if they are structs in the
first place), then the container cannot be an array.