Discard the program configuration object when GDB startup fails.
Before GDB is started, the program configuration is allocated, but only
read partially, so that the debugger command is known. If GDB cannot be
started later, the incomplete configration container remains. The user sees
an error message, and, after dismissing it, can exit KDbg. At this time,
the incomplete configuration is saved in the destructor, erasing all
settings that would have been present. Discard the incomplete configuration
after a failed GDB startup.

This also fixes an assertion failure in openProgramConfig() if after the
failed startup a successful startup occurs.