Parse newer formats of the table of shared libraries.
The list of shared libraries looks different from when the parser was
implemented. We can now have an indicator "(*)! after "Yes" (which is in
the column "Symb Read" that tells whether debugging information was found.
This means that a line can have 4 tokens instead of just 3 before the file
name. The indicator and the spaces following it now become part of the
shared library name. This does not hurt in practice because the names are
only used to find out which type tables apply, and these use liberal
regular expressions that look only at the end of the name.

Nevertheless, adjust the parser to recognize the new format. Pick out the
last non-space part as the file name from each line. This does assume that
there are no space characters in shared library names.

This also accounts for another kind of lines that do not carry addresses.
If the "(*)" indicator is present, the old parser happened to do the right
thing, but if it is not present, it took the entire next line as the name
of a shared library.