Do not skip duplicated local variables.
When C code reuses names of variables in nested blocks, the inner name
hides the outer name. GDB prints local variables starting with the
innermost block. Function KDebugger::parseLocals() has logic in place
to handle this case such that the display shows the value of the innermost
variable under the unadorned name and the hidden values under names
decorated with a counter. This code exists since the beginning of KDbg.

But then in commit 01720cbaf6ce the parser was changed to remove the
duplicated variables. As a consequence, only the innermost name remains
visible. The reason for this change is unknown. Revert it, because we
prefer to see the values of the hidden variables. Add a test case.