Retrieve widget from the tree item instead of keeping a separate pointer.
Command items that update a particular entry in the Locals and Watches
windows need to reference that item. When the GDB output is handled, the
updated value must be moved to the respective window. This requires the
ExprWnd instance. Since the items themselves know in which tree widget they
live, it is not necessary to keep a separate pointer. Remove the pointer
and ask the tree item for the widget when needed.